+++ src/PixelShader8.cpp
@@ -834,7 +834,7 @@ int CPixelShader8::Get_FX_Params_Number()
 	else if (wcscmp(m_FX_Name, L"SpatialDistorsion") == 0) NumberParams = 3;
 	else if (wcscmp(m_FX_Name, L"Displacement") == 0) NumberParams = 3;
 	else if (wcscmp(m_FX_Name, L"Radar") == 0) NumberParams = 5;
-	else if (wcscmp(m_FX_Name, L"Sharpen") == 0) NumberParams = 1;
+	else if (wcscmp(m_FX_Name, L"Sharpen1") == 0) NumberParams = 1;
 	else NumberParams = 0;
 
 	return NumberParams;
@@ -990,10 +990,10 @@ void  CPixelShader8::Display_FX_Param1(char* outParam, int outParamSize, float v
 			float Speed = ParamAdjust(value, 0.0f, 10.0f);
 			sprintf_s(outParam, outParamSize, "%.2f (Speed)", Speed);
 		}
-		else if (wcscmp(m_FX_Name, L"Sharpen") == 0)
+		else if (wcscmp(m_FX_Name, L"Sharpen1") == 0)
 		{
-			float Sharpness = ParamAdjust(value, 0.0f, 3.0f);
-			sprintf_s(outParam, outParamSize, "%.2f (Sharpness)", Sharpness);
+			float Amount = ParamAdjust(value, 0.0f, 20.0f);
+			sprintf_s(outParam, outParamSize, "%.2f (Amount)", Amount);
 			}
 		else
 		{